# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Grocery API Doc",
    "version": "1.0.0",
    "description": "This is API documentation for Grocery API created as an assignment for QP"
  },
  "paths": {
    "/api/admin/products": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get All products for Admin",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/admin/product/create": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Create a product",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "price",
                  "stock"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the product"
                  },
                  "description": {
                    "type": "string",
                    "description": "The description of the product"
                  },
                  "price": {
                    "type": "number",
                    "description": "The price of the product"
                  },
                  "stock": {
                    "type": "number",
                    "description": "stock of the product"
                  },
                  "active": {
                    "type": "boolean",
                    "description": null
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/admin/product/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get Product by Id for Admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the Product to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      },
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Udpate Product by Id for Admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the Product to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "summary": "Delete Product by Id for Admin",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the Product to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Retrieve an example message",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Adesh Tamrakar"
                  },
                  "email": {
                    "type": "string",
                    "example": "adesh.t111@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "StrongPassword123#"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Retrieve an example message",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "adesh.t111@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "StrongPassword123#"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, JWT token generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "role": {
                              "type": "string"
                            },
                            "createdAt": {
                              "type": "string"
                            },
                            "updatedAt": {
                              "type": "string"
                            }
                          }
                        },
                        "token": {
                          "type": "string",
                          "description": "JWT Token for authentication"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get Alls Products info",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/users/product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Get Product info by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the Product to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/users/orders": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Get all orders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/users/order": {
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Create an order",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orderItems"
                ],
                "properties": {
                  "orderItems": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productId": {
                          "type": "string",
                          "description": "The Id the product"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Quantity of the product"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/users/order/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Get Order info by Id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the Order to retrieve",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/users/profile": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get Profile info of logged-in User",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    },
    "/api/users/{id}": {
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Update user own profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The ID of the user to retrieve",
            "schema": {
              "type": "string",
              "example": "b6e3e5a9-0812-49f7-bd5f-57043fc1333a"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Auth",
      "description": "All authentication APIs"
    },
    {
      "name": "User",
      "description": "All user APIs"
    },
    {
      "name": "Admin",
      "description": "All user APIs"
    },
    {
      "name": "Order",
      "description": "Order APIs for users"
    },
    {
      "name": "Product",
      "description": "Product APIs for users"
    }
  ]
}